# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API Document",
        "title": "API for POS",
        "version": "v1"
    },
    "schemes": "HTTP",
    "basePath": "/",
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "name": "X-POS-User-Token",
                "in": "header"
            }
        }
    },
...
    "paths": {
        "/users/signup": {
            "post": {
                "tags": [
                    "user API"
                  ],
                "summary": "Create new user",
                "parameters": [{
                    "in": "formData",
                    "name": "email",
                    "type": "string",
                    "description": "Email address of new user",
                    "required": true
                }, {
                    "in": "formData",
                    "name": "password",
                    "type": "string",
                    "description": "Password of new user",
                    "required": true
                }, {
                    "in": "formData",
                    "name": "name",
                    "type": "string",
                    "description": "Name of new user",
                    "required": true
                }, {
                    "in": "formData",
                    "name": "cellnumber",
                    "type": "string",
                    "description": "Phone number of new user",
                    "required": true
                }],
                "responses": {
                    "200":{
                        "description": "User successfully created"
                    },
                    "201":{
                        "description": "The email already exist"
                    },
                    "405":{
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "user API"
                  ],
                "summary": "User signs in",
                "parameters": [{
                    "in": "formData",
                    "name": "email",
                    "type": "string",
                    "description": "Email address of user",
                    "required": true
                }, {
                    "in": "formData",
                    "name": "password",
                    "type": "string",
                    "description": "Password of user",
                    "required": true
                }],
                "responses": {
                    "200":{
                        "description": "User successfully logged in"
                    },
                    "201":{
                        "description": "Username or password is incorrect"
                    },
                    "405":{
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    }
}